circular deps fixed
