+++ modules/@angular/forms/src/directives/reactive_directives/form_group_directive.ts
@@ -20,7 +20,7 @@ import {NgControl} from '../ng_control';
 import {ReactiveErrors} from '../reactive_errors';
 import {composeAsyncValidators, composeValidators, setUpControl, setUpFormContainer} from '../shared';
 
-import {FormArrayName} from './form_array_name';
+import {FormArrayName} from './form_group_name';
 import {FormGroupName} from './form_group_name';
 
 export const formDirectiveProvider: any = {
